pyproject.toml
[project]
name = "budgie"
version = "0.1.0"
description = "Personal household budget management web application"
authors = [
    {name = "David"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "sqlalchemy[asyncio] (>=2.0.47,<3.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "ofxtools (>=0.9.5,<0.10.0)",
    "quiffen (>=4.0.1,<5.0.0) ; python_version < \"3.14\""
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "ruff (>=0.15.4,<0.16.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pandas-stubs (>=3.0.0.260204,<4.0.0.0)"
]

# ── Tool configuration ─────────────────────────────────────────────

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.ruff]
target-version = "py312"
line-length = 88
src = ["budgie", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "A",    # flake8-builtins
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
    "D",    # pydocstyle
    "ANN",  # flake8-annotations
]
ignore = [
    "D100",   # Missing docstring in public module (handled per-file)
    "D104",   # Missing docstring in public package
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D103", "ANN"]  # No docstrings or annotations required in tests
"alembic/**/*.py" = ["D", "ANN", "E501", "W"]  # Auto-generated migration files

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "ofxtools.*",
    "quiffen.*",
    "aiosqlite.*",
    "jose.*",
    "bcrypt.*",
]
ignore_missing_imports = true
